refactor: split sharpe.py into a topical sub-package (#256)
Break the monolithic 1182-LOC src/jsharpe/sharpe.py (26 functions) into a
src/jsharpe/sharpe/ package grouping related routines:

- linalg.py: ppoints, robust_covariance_inverse, minimum_variance_weights
- clustering.py: effective_rank, _silhouette_samples, number_of_clusters
- quadrature.py: make_expectation_gh, E_under_normal, moments_Mk
- psr.py: sharpe_ratio_variance, variance/expected max, track record,
  critical_sharpe_ratio, probabilistic_sharpe_ratio, sharpe_ratio_power
- corrections.py: Bonferroni/Šidák/Holm, FDR_critical_value, control_for_FDR,
  pFDR, oFDR
- generators.py: data-generation helpers + autocorrelation

The full public API (and moments_Mk, used by tests) is re-exported unchanged
through sharpe/__init__.py, so 'from jsharpe import ...',
'from jsharpe.sharpe import ...' and the __all__ contract resolve identically.
No single module exceeds ~290 LOC.

Closes #256.
